requirements.txt
# Core dependencies
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings>=2.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4

# Monitoring
prometheus-client>=0.11.0
prometheus-fastapi-instrumentator>=5.7.1

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.23.0

# Data processing
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
textblob==0.17.1
emoji==2.8.0
textstat==0.7.3

# Utilities
python-dotenv==1.0.0
apify-client==1.6.1
openai==1.3.7
redis>=4.5.0
tqdm==4.66.2

# Dashboard
streamlit==1.29.0
plotly==5.18.0 